+++ src/vivarium_public_health/exposure/__init__.py
@@ -1,0 +1,2 @@
+from .effect import ExposureEffect
+from .exposure import Exposure