+++ CHANGELOG.rst
@@ -1,3 +1,10 @@
+**4.3.0 - TBD/TBD/25**
+
+  - Refactor to create two new abstract classes, Exposure and ExposureEffect
+  - Add Intervention and InterventionEffect classes
+  - Refactor to pass Exposure component to RiskExposureDistribution classes
+  - Update exposure category names for DichotomousRiskExposureDistribution
+  
 **4.2.5 - 07/18/25**
 
   - Hush FutureWarnings: add include_groups=False to apply call in StratifiedObservation